requirements.txt
# ── Already installed on this machine (torch+cuda, transformers, pandas, etc.) ──
# These are listed for documentation; pip will skip if already satisfied.

# ── Google Calendar API (optional - install for automatic sync) ──────────────
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Core parsing
python-dateutil==2.9.0.post0

# Config
pyyaml>=6.0

# Data
pandas>=2.0
numpy>=1.24

# Git
GitPython>=3.1

# HTTP (for Ollama)
httpx>=0.25

# Embeddings
sentence-transformers>=3.0

# Vector DB (in-process, no Docker)
qdrant-client>=1.9

# UI
streamlit>=1.40
altair>=5.0
watchdog>=4.0

# CLI
click>=8.0
rich>=13.0

# Logging
loguru>=0.7

# Streaming JSON (for large Google Timeline files)
ijson>=3.2

# Validation
pydantic>=2.0

# Progress
tqdm>=4.60
